migration from js to ts of youtube bot
